pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["EnterNick <Ravil.Mirgayazov@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"
pillow = "^11.0.0"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
ruff = "^0.8.3"
django-channels = "^0.7.0"
daphne = "^4.1.2"
channels = "^4.2.0"
channels-redis = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "W", "C901", "B"]
ignore = []

[tool.ruff.format]
quote-style = "single"

indent-style = "space"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"